requirements.txt
flask
flask-cors
speedtest-cli
ping3
requests
